Extract `initializeLocalizationFSWatcher` logic out
