+++ src/extension.ts
@@ -2,12 +2,14 @@
 // The module `vscode` contains the VS Code extensibility API
 import * as vscode from 'vscode';
 
+// Initializers
+import * as initialize from './initializers';
+
 // Commands
 import { commands } from './commands';
 
 // Providers
 import { providers } from './providers';
-import { buildLocalizationReference } from './library/bg3/buildLocalizationReference';
 
 
 // This method is called when your extension is activated
@@ -18,28 +20,8 @@ export async function activate(context: vscode.ExtensionContext) {
 	// Show Information Message when the extension is activated
 	vscode.window.showInformationMessage("BG3 Modding Extension Activated!");
 
-	// Create a file-watcher for the xml localization files
-	const localizationXMLFileWatcher = vscode.workspace.createFileSystemWatcher("**/Localization/**/*.xml");
-
-	// Build localization references when a new xml is created
-	localizationXMLFileWatcher.onDidCreate(e => {
-		buildLocalizationReference([e]);
-	});
-
-	// Rebuild localization references when a xml is changed
-	// ? Performance concerns?
-	localizationXMLFileWatcher.onDidChange(e => {
-		buildLocalizationReference([e]);
-	});
-
-	// Remove localization references when a xml is deleted
-	localizationXMLFileWatcher.onDidDelete(e => {
-		/**
-		 * // TODO: Remove references. May also have to dispose off the registered hover provider
-		 * ? (if I decide to register separate providers for each file)
-		 */
-		throw new Error("TODO: Implementation pending!");
-	});
+	// Initialize Localization FS Watcher
+	initialize.localizationFSWatcher();
 
 	// Register all the commands and providers, and subscribe to their disposables
 	context.subscriptions.push(
